Extract unwind_without_debuginfo
